Reorganise C includes
