+++ src/kernel/json-parse.c
@@ -3,9 +3,9 @@
 
 #include <stdio.h>
 
-#include "./gc.h"
-#include "./json.h"
-#include "./string.h"
+#include "gc.h"
+#include "json.h"
+#include "string.h"
 
 Custom Json_Value_null = {.header = HEADER_CUSTOM(0), .ctor = JSON_VALUE_NULL};
