+++ src/kernel/kernel.h
@@ -4,14 +4,14 @@
 #define EMSCRIPTEN_KEEPALIVE
 #endif
 
-#include "./basics.h"
-#include "./char.h"
-#include "./debug.h"
-#include "./elm.h"
-#include "./gc.h"
-#include "./json.h"
-#include "./list.h"
-#include "./string.h"
-#include "./types.h"
-#include "./utils.h"
-#include "./wrapper/wrapper.h"
+#include "basics.h"
+#include "char.h"
+#include "debug.h"
+#include "elm.h"
+#include "gc.h"
+#include "json.h"
+#include "list.h"
+#include "string.h"
+#include "types.h"
+#include "utils.h"
+#include "wrapper/wrapper.h"
