+++ src/test/char_test.c
@@ -1,5 +1,5 @@
 #include "../kernel/kernel.h"
-#include "./test.h"
+#include "test.h"
 
 /*
 https://en.wikipedia.org/wiki/UTF-16#Examples
