+++ src/test/gc/replay/apply_alloc_failed.c
@@ -1,3 +1,5 @@
+#include "../../gc_test.h"
+#include "../../test.h"
 #include "../replay_test.h"
 
 char* test_replay_apply_alloc_failed() {
@@ -21,13 +23,13 @@ char* test_replay_apply_alloc_failed() {
       gc_state.heap.end - gc_state.next_alloc,
       words_to_leave);
 
-  void* result = Utils_apply(&Utils_equal, 2, (void* []){&Unit, &Unit});
+  void* result = Utils_apply(&Utils_equal, 2, (void*[]){&Unit, &Unit});
   mu_expect_equal("Heap should overflow on making a function call", result, pGcFull);
 
   GC_collect_full();
   GC_prep_replay();
 
-  result = Utils_apply(&Utils_equal, 2, (void* []){&Unit, &Unit});
+  result = Utils_apply(&Utils_equal, 2, (void*[]){&Unit, &Unit});
 
   void* h = gc_state.heap.start;
