+++ src/test/gc/replay_test.h
@@ -1,16 +1,8 @@
 #ifndef REPLAY_TEST_H
 #define REPLAY_TEST_H
 
-#include <stdbool.h>
-#include <stdio.h>
-#include <string.h>
-#include "../../kernel/debug.h"
 #include "../../kernel/gc-internals.h"
-#include "../../kernel/gc.h"
-#include "../../kernel/types.h"
-#include "../../kernel/utils.h"
-#include "../gc_test.h"
-#include "../test.h"
+#include "../../kernel/kernel.h"
 
 char* replay_scenario_tests();
 Closure mock_func;
