+++ src/test/gc/stackmap_mark_test.c
@@ -3,11 +3,8 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "../../kernel/debug.h"
 #include "../../kernel/gc-internals.h"
-#include "../../kernel/gc.h"
-#include "../../kernel/list.h"
-#include "../../kernel/types.h"
+#include "../../kernel/kernel.h"
 #include "../gc_test.h"
 #include "../test.h"
