+++ src/test/json/json_decoder_test.c
@@ -1,10 +1,11 @@
 #include <stddef.h>
 
 #include "../../kernel/kernel.h"
-#include "./../test.h"
+#include "../../kernel/json.h"
+#include "../test.h"
 
 ElmValue* parse_json(ElmString16* json);
-void* Json_wrap(void*);
+void* wrap(void*);
 
 //=================================================================================================
 // Test helpers
@@ -78,7 +79,7 @@ void* test_Json_decode_invalidJson() {
           2,
           ((void*[]){
               create_string("This is not valid JSON!"),
-              Json_wrap(create_string(json)),
+              wrap(create_string(json)),
           }))));
 
   return NULL;
@@ -123,7 +124,7 @@ void* test_Json_decodeString() {
 void* test_Json_decodeValue() {
   test_decode_ok(&Json_decodeValue,
       "[1,2,3]",
-      Json_wrap(NEW_CUSTOM(JSON_VALUE_ARRAY,
+      wrap(NEW_CUSTOM(JSON_VALUE_ARRAY,
           3,
           ((void*[]){
               NEW_ELM_FLOAT(1),
