+++ src/test/wrapper/src/mock-app.c
@@ -5,13 +5,9 @@
 #define EMSCRIPTEN_KEEPALIVE
 #include <stdio.h>
 #endif
-#include "../../../kernel/basics.h"
-#include "../../../kernel/debug.h"
+
 #include "../../../kernel/gc-internals.h"
-#include "../../../kernel/gc.h"
-#include "../../../kernel/types.h"
-#include "../../../kernel/utils.h"
-#include "../../../kernel/wrapper.h"
+#include "../../../kernel/kernel.h"
 
 /* ---------------------------------------------------------
